backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
spacy==3.7.2
transformers==4.36.0
torch>=2.2.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
nltk>=3.8.0
beautifulsoup4>=4.12.0
aiofiles>=23.2.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
